Introducing test.sh
test.sh will help with the development cycle of this project. I'll use
it to package the project into an excutable jar file and run some
command on it in an example project, which can be found in
src/test/example_project.

As I work on fractoring the codebase, I'll work on the package feature
next.